List.of() gives required spliterator but requires JDK9 or above
