+++ unit-tests/shared/src/test/scala/org/scalanative/testsuite/javalib/util/stream/IntStreamTest.scala
@@ -12,8 +12,9 @@ package org.scalanative.testsuite.javalib.util.stream
 import java.{lang => jl}
 
 import java.{util => ju}
-import java.util.{Arrays, List}
-import java.util.{OptionalInt, IntSummaryStatistics}
+import java.util.Arrays
+import java.util.IntSummaryStatistics
+import java.util.OptionalInt
 import java.util.{Spliterator, Spliterators}
 
 import java.util.concurrent.{CountDownLatch, TimeUnit}
@@ -887,7 +888,7 @@ class IntStreamTest {
 
   @Test def intStreamFindAny_True(): Unit = {
     val s = IntStream.of(0, 11, 22, 33)
-    val acceptableValues = List.of(0, 11, 22, 33)
+    val acceptableValues = List(0, 11, 22, 33)
 
     val optional = s.findAny()
 
@@ -1477,32 +1478,6 @@ class IntStreamTest {
       expectedValue,
       iter.nextInt()
     )
-  }
-
-  // Issue 4007
-  @Test def intStreamSkip_GivesDownstreamAccurateExpectedSize(): Unit = {
-    /* Tests for Issue 4007 require a SIZED spliterator with a tryAdvance()
-     * which does not change the exactSize() after traversal begins.
-     *
-     * This Test is fragile in that it uses intimate knowledge of
-     * Scala Native internal implementations to provide such a spliterator. If
-     * those implements change, this Test may end up succeeding but
-     * not exercising the Issue 4007 path.
-     *
-     * List.of() followed by mapToInt() provides a suitable spliterator.
-     * IntStream.of() by itself does not provoke the defect, its exactSize()
-     * bookkeeping is too good.
-     */
-
-    val srcData = List.of(111, 222, 333, 444, 555, 666, 777)
-    val s = srcData.stream()
-    val is = s.mapToInt(e => e)
-
-    val skipSize = 4
-    val expectedSize = srcData.size() - skipSize
-    val resultSize = is.skip(skipSize).toArray().size
-
-    assertEquals("expectedSize", expectedSize, resultSize)
   }
 
   @Test def intStreamSorted(): Unit = {
