+++ unit-tests/shared/src/test/scala/org/scalanative/testsuite/javalib/util/stream/StreamTest.scala
@@ -1374,28 +1374,6 @@ class StreamTest {
     assertEquals("unexpected first value: ", expectedValue, iter.next())
   }
 
-  // Issue 4007
-  @Test def streamSkip_GivesDownstreamAccurateExpectedSize(): Unit = {
-    /* Use List.of() in Tests for Issue 4007. That Issue requires
-     * a SIZED spliterator with a tryAdvance() which does not
-     * change the exactSize() after traversal begins. List.of() provides such.
-     *
-     * Stream.of() does not exercise the condition being tested.
-     * The implementation path it uses does bookkeeping to provide
-     * an accurate exactSize() after traversal begins. That work is
-     * allowed but not required by the JDK.
-     */
-
-    val srcData = List.of("R", "S", "T", "U", "V", "X", "Y", "Z")
-    val s = srcData.stream()
-
-    val skipSize = 4
-    val expectedSize = srcData.size() - skipSize
-    val resultSize = s.skip(skipSize).toList.size()
-
-    assertEquals("expectedSize", expectedSize, resultSize)
-  }
-
   @Test def streamSorted(): Unit = {
     val nElements = 8
     val wild = new ArrayList[String](nElements)
